Refactor UI components and styles: update image source in SignupPage, modify Button imports, enhance LoginForm and SignupForm button styles, and rename CSS variables for clarity.
