+++ app/(pages)/(dashboard)/settings/page.jsx
@@ -3,7 +3,7 @@ import React from 'react'
 const page = () => {
   return (
     <div>
-      
+      hi
     </div>
   )
 }
