+++ styles/globals.css
@@ -124,9 +124,8 @@
 }
 
 @theme{
-  --color-primary: #092601;
+  --color-basic: #092601;
   --color-secondary: #265902;
   --color-accent: #265902;
   --color-highlight: #98BF0A;
-  --color-light: #F4F3F8;
 }
